Move OpenResearcher placeholder cleanup to standardization
