+++ datasets/openresearcher/raw_to_atif.py
@@ -5,10 +5,8 @@ import json
 import sys
 from typing import Any
 
-from schema.atif import Step
 from scripts.raw_to_atif_common import (
     dataset_name_from_script,
-    renumber_steps,
     text_from_content,
     trajectories_from_input,
 )
@@ -18,15 +16,6 @@ BROWSER_TOOL_NAMES = {
     "browser.open": "browser.open",
     "browser.find": "browser.find",
 }
-
-
-def is_placeholder_system_message(step: Step) -> bool:
-    return (
-        step.source == "system"
-        and not step.tool_calls
-        and step.observation is None
-        and text_from_content(step.message).strip() in {"", "None"}
-    )
 
 
 def add_browser_tool_names(record: Any) -> Any:
@@ -52,9 +41,6 @@ def main(script_file: str) -> None:
     dataset_name = dataset_name_from_script(script_file)
     records = (add_browser_tool_names(json.loads(line)) for line in sys.stdin if line.strip())
     for trajectory in trajectories_from_input(records, dataset_name):
-        trajectory.steps = renumber_steps(
-            [step for step in trajectory.steps if not is_placeholder_system_message(step)]
-        )
         print(trajectory.model_dump_json(exclude_none=True))
